+++ multimodal/tests/unittests/others/test_matcher.py
@@ -11,8 +11,8 @@ from autogluon.multimodal import MultiModalPredictor
 from autogluon.multimodal.constants import BINARY, MULTICLASS, QUERY, RESPONSE, UNIFORM_SOUP
 from autogluon.multimodal.utils import convert_data_for_ranking, semantic_search
 
-from .unittest_datasets import Flickr30kDataset, IDChangeDetectionDataset
-from .utils import get_home_dir
+from ..utils.unittest_datasets import Flickr30kDataset, IDChangeDetectionDataset
+from ..utils.utils import get_home_dir
 
 ALL_DATASETS = {
     "id_change_detection": IDChangeDetectionDataset,
