+++ multimodal/tests/unittests/others_2/test_data_process_image.py
@@ -4,7 +4,7 @@ import pytest
 from autogluon.multimodal.data.process_image import ImageProcessor
 from autogluon.multimodal.models import TimmAutoModelForImagePrediction
 
-from ..others.unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
+from ..utils.unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
 
 ALL_DATASETS = {
     "petfinder": PetFinderDataset,
