+++ multimodal/tests/unittests/others_2/test_predictor_advanced.py
@@ -9,7 +9,7 @@ from torch import Tensor
 from autogluon.multimodal import MultiModalPredictor
 from autogluon.multimodal.constants import BIT_FIT, IA3, IA3_BIAS, IA3_LORA, LORA_BIAS, LORA_NORM, NORM_FIT
 
-from ..others.unittest_datasets import AmazonReviewSentimentCrossLingualDataset
+from ..utils.unittest_datasets import AmazonReviewSentimentCrossLingualDataset
 
 
 def _is_lazy_weight_tensor(p: Tensor) -> bool:
